+++ src/actions/sortListActions.js
@@ -1,5 +1,5 @@
 import { SET_SORT } from './actionTypes';
-import { storage } from '../components/Dashboard/SortList/utils';
+import { storage } from '../components/Dashboard/SortList/sortUtils';
 
 const setSort = (key, field, direction) => dispatch => {
   storage.set(key, {
